requirements.txt
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
PyYAML==6.0.3
requests==2.32.5
tzlocal==5.3.1
urllib3==2.6.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
